apps/slack/rag.py
import os
from langchain_openai import ChatOpenAI, OpenAIEmbeddings
from langchain_community.vectorstores import Qdrant
from langchain.chains import RetrievalQA
from langchain.schema import Document

class RAGService:
    def __init__(self):
        self.openai_api_key = os.environ.get("OPENAI_API_KEY")
        self.qdrant_url = os.environ.get("QDRANT_URL", "http://localhost:6333")
        
        # Mock initialization if keys are missing
        if self.openai_api_key:
            self.embeddings = OpenAIEmbeddings()
            self.llm = ChatOpenAI(model_name="gpt-4-turbo")
            # Connect to Qdrant here
            # self.vectorstore = Qdrant(client=..., collection_name="safespace", embeddings=self.embeddings)
        else:
            print("Warning: OPENAI_API_KEY not set. RAG service will be mocked.")

    def query(self, question: str) -> str:
        if not self.openai_api_key:
            return "I'm sorry, I can't answer that right now (RAG service not configured)."
            
        # Mock retrieval
        # qa = RetrievalQA.from_chain_type(llm=self.llm, chain_type="stuff", retriever=self.vectorstore.as_retriever())
        # return qa.run(question)
        
        return f"Here is some information about '{question}' based on our company policies..."

rag_service = RAGService()
