apps/slack/requirements.txt
slack_bolt>=1.18.0
slack_sdk>=3.26.0
langchain>=0.1.0
langchain-openai>=0.0.5
qdrant-client>=1.7.0
cryptography>=42.0.0
python-dotenv>=1.0.0
